orchestration-service/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
passlib
bcrypt==4.0.1
python-jose[cryptography]
httpx
python-dotenv